move link fn to links.ts
